% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-front-matter.R
\name{clean_front_matter}
\alias{clean_front_matter}
\title{Clean up intermediate files created by \code{comp_front_*}}
\usage{
clean_front_matter(files)
}
\description{
Clean up intermediate files created by \code{comp_front_*}
}
\keyword{internal}
