% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-front-matter.R
\name{comp_front}
\alias{comp_front}
\title{Compile front matter to PDF}
\usage{
comp_front(use_docker = FALSE, template = c("ntu"))
}
\arguments{
\item{use_docker}{Logical. Whether to use tinytex docker
to compile PDF.}

\item{template}{String. The template used. Defaults to
\code{ntu}.}
}
\description{
To use this function, one needs to be in the
\emph{root directory} of the bookdown project.
}
\examples{
\dontrun{
setwd("~/my-thesis")
comp_front(template = 'ntu')
}
}
