% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-front-matter-ntu.R
\name{comp_front_ntu}
\alias{comp_front_ntu}
\title{Compile front matter to PDF}
\usage{
comp_front_ntu()
}
\description{
To use this function, one needs to be in the
\emph{root directory} of the bookdown project.
}
\examples{
\dontrun{
setwd("~/my-thesis")
comp_front(template = 'ntu')
}
}
\keyword{internal}
