% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-front-matter.R
\name{construct_front_rmd}
\alias{construct_front_rmd}
\title{Construct 'front_matter.rmd'}
\usage{
construct_front_rmd(person_info = "../person-info.yml")
}
\arguments{
\item{person_info}{File path to \code{person-info.yml}.}
}
\description{
Construct \code{front_matter.rmd} from \code{front_matter/output.yml}
and \code{person-info} in root.
}
\keyword{internal}
