% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-tinytex.R
\name{docker_xelatex}
\alias{docker_xelatex}
\title{Wrapper of Docker `liao961120/tinytex` to Compile PDF}
\usage{
docker_xelatex(fpath)
}
\arguments{
\item{fpath}{String. The path to \code{.tex} file.}
}
\description{
Wrapper of Docker `liao961120/tinytex` to Compile PDF
}
\references{
\url{https://hub.docker.com/r/liao961120/tinytex}
}
