% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_font.R
\name{donwload_font2header}
\alias{donwload_font2header}
\title{Download font dependencies to the directory where header.tex is}
\usage{
donwload_font2header()
}
\description{
Make sure this function is called after 1) importing Rmd template
and 2) setting the current working directory to that Rmd
template folder.
}
\keyword{internal}
