% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_font.R
\name{donwload_font_all}
\alias{donwload_font_all}
\title{Download traditional Chinese font dependencies Rmd PDF output}
\usage{
donwload_font_all(lookfor = "template.tex")
}
\description{
Make sure this function is called after 1) importing Rmd template
and 2) setting the current working directory to that Rmd
template folder.
}
