% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-project.R
\name{init_proj}
\alias{init_proj}
\title{Initialize current directory as a bookdown project}
\usage{
init_proj(dir = ".", download_font = TRUE)
}
\arguments{
\item{dir}{The path to the dircetory to initialize as
a project.}

\item{download_font}{Logical. Whether to download \code{kaiu.ttf}
for CJK mainfont & \code{NotoSansMonoCJKtc.otf} for CJK mono font.
Defaults to \code{TRUE}.}
}
\description{
Initialize current directory as a bookdown project
}
\examples{
\dontrun{
setwd("~/my-thesis")
init_proj()
}

}
