% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_book.R
\name{tex_book}
\alias{tex_book}
\title{Modified Version of `bookdown::pdf_book`}
\usage{
tex_book(rename = "tex_book.tex", toc = TRUE, number_sections = TRUE,
  fig_caption = TRUE, pandoc_args = NULL, ...,
  base_format = rmarkdown::pdf_document, toc_unnumbered = TRUE,
  toc_appendix = FALSE, toc_bib = FALSE, quote_footer = NULL,
  highlight_bw = FALSE)
}
\arguments{
\item{toc}{See bookdown::pdf_book}

\item{number_sections}{See bookdown::pdf_book}

\item{fig_caption}{See bookdown::pdf_book}

\item{pandoc_args}{See bookdown::pdf_book}
}
\description{
Modified Version of `bookdown::pdf_book`
}
