% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2pdf.R
\name{word2pdf}
\alias{word2pdf}
\title{Compile Word into PDF}
\usage{
word2pdf(infile = "word.docx", outfile = "word.pdf")
}
\arguments{
\item{infile}{String. File path to the input word document.
Must be in \code{.docx} format.}

\item{outfile}{String. File path of the output pdf document
to be generated. Must be in \code{.pdf} format.}
}
\description{
Pandoc Wrapper function
}
\keyword{internal}
