% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-front-matter.R
\name{comp_front}
\alias{comp_front}
\title{Compile front matter to PDF}
\usage{
comp_front()
}
\description{
To use this function, one needs to be in the
\emph{root directory} of the bookdown project.
}
