% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-template.R
\name{import_template}
\alias{import_template}
\title{Import Thesis Template}
\usage{
import_template(name = "My_Thesis")
}
\arguments{
\item{name}{The name of the thesis project to create.}
}
\description{
\code{import_template} creates a thesis template. It is a
  wrapper of \code{\link{rmarkdown::draft}}.
}
\examples{
\dontrun{
import_template("my-thesis")
}

}
