% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-project.R
\name{init_proj}
\alias{init_proj}
\title{Initialize current directory as a bookdown project}
\usage{
init_proj()
}
\description{
Initialize current directory as a bookdown project
}
\examples{
init_proj()
}
